requirements.txt
chromadb>=0.5.0
deepeval>=3.8.4
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-chroma>=0.1.0
langchain-ollama>=0.1.0
ollama>=0.6.0
streamlit>=1.32.0
pypdf>=4.0.0
sentence-transformers>=3.0.0
llama-cpp-python>=0.2.78
langchain-huggingface>=0.0.3
arize-phoenix>=12.0.0
arize-phoenix-otel>=0.14.0
